Add files via upload
In this project, I have discussed and elaborated on how an R programmer can perform all of the operations of SQL in an R programming setting without having to leave his/her comfort zone. For this project, I have connected to a cloud database on AWS. The queries have been built from the basic levels to more advanced queries.